Vismaya

+++ RetailStore/Model/Order.cs
@@ -4,6 +4,6 @@ public class Order: DomainAudit
 {
     public int  Id { get; set; }
     public int CustomerId { get; set; }
-    public int TotalAmount { get; set; }
+    public double TotalAmount { get; set; }
     public virtual Customer Customer { get; set; }
 }
